# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super  ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
        "/api/addcollaborator": {
            "post": {
                "summary": "Add Colaborator to a specific Note",
                "description": "Add Colaborator a to specific Note",
                "operationId": "37eee3800ca4c13011d3c47bfa02aa70",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "note_id",
                                    "email"
                                ],
                                "properties": {
                                    "note_id": {
                                        "type": "integer"
                                    },
                                    "email": {
                                        "type": "email"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Collaborator Created Sucessfully"
                    },
                    "202": {
                        "description": "Collab Not Added"
                    },
                    "401": {
                        "description": "Invalid Authorization Token"
                    },
                    "404": {
                        "description": "Not a Registered Email"
                    },
                    "409": {
                        "description": "Collaborator Already Created"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/updatecollaborator": {
            "post": {
                "summary": "update Colaborator",
                "description": " Update Colaborator Note ",
                "operationId": "07b0a24e50355e5ebe1b1a0f22309633",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "note_id",
                                    "updated_title",
                                    "updated_description"
                                ],
                                "properties": {
                                    "note_id": {
                                        "type": "integer"
                                    },
                                    "updated_title": {
                                        "type": "string"
                                    },
                                    "updated_description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Collaborator updated Sucessfully"
                    },
                    "401": {
                        "description": "Invalid Authorization Token"
                    },
                    "404": {
                        "description": "Not a Registered Email"
                    },
                    "409": {
                        "description": "Collaborator Already Created"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/deletecollaborator": {
            "delete": {
                "summary": "delete Colaborator",
                "description": " Delete Colaborator Note ",
                "operationId": "c7141baa24374c94b9202e2c8a9ee5a3",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "note_id",
                                    "email"
                                ],
                                "properties": {
                                    "note_id": {
                                        "type": "integer"
                                    },
                                    "email": {
                                        "type": "email"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Collaborator deleted Sucessfully"
                    },
                    "401": {
                        "description": "Invalid Authorization Token"
                    },
                    "404": {
                        "description": "Not a Registered Email"
                    },
                    "409": {
                        "description": "Collaborator Already Created"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "description": "laravel bearer authentication",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}